Move getFavoriteCommunities to a central place
