+++ src/App.tsx
@@ -16,8 +16,7 @@ import "@ionic/react/css/text-transformation.css";
 import "@ionic/react/css/flex-utils.css";
 import "@ionic/react/css/display.css";
 
-import { Provider } from "react-redux";
-import store from "./store";
+import { StoreProvider } from "./store";
 import { isInstalled } from "./helpers/device";
 import TabbedRoutes from "./TabbedRoutes";
 import Auth from "./Auth";
@@ -38,7 +37,7 @@ setupIonicReact({
 export default function App() {
   return (
     <AppContextProvider>
-      <Provider store={store}>
+      <StoreProvider>
         <GlobalStyles>
           <BeforeInstallPromptProvider>
             <UpdateContextProvider>
@@ -54,7 +53,7 @@ export default function App() {
             </UpdateContextProvider>
           </BeforeInstallPromptProvider>
         </GlobalStyles>
-      </Provider>
+      </StoreProvider>
     </AppContextProvider>
   );
 }
